Palindrome
Determine if the number is a palindrome.
By modding(% 10) the input, we can extract the right most digit.
Then multiply this digit by 10, and divide the input by 10.
Repeat the modding (%10) step again, this time add the new right digit to the reverseX variable.
Now we will be able to keep adding more digits to reverseX.
Lastly, compare reverseX and input. If they are the same then true if they are not return false.