Reverse left side of arr, sum every 2 element on right side of arr
Given an array of integers of any size, n ≥ 1, write an algorithm as a pseudo code (not a
program!) that would reverse every two consecutive elements of the left half of the array (i.e.
reverse elements at index 0 &1, at index 2 & 3, etc.). Additionally, for the right half of the
array, the algorithm must change the second element of every two consecutive elements to
have to the sum of the two elements. For instance, if the right half starts at index 14, and the
values at index 14 & 15 are 72 and 11, then the value at index 15 is changed to 83. If the given
array is of an odd size, then the exact middle element should not be manipulated by the
algorithm in any way